requirements.txt
matplotlib
Mako
pypika
asyncpg
pytz
python-dateutil
python-dotenv
fastapi
uvicorn[standard]

